refactor: extract IncludeLoader class (#25)
